eShopLite: Split EF model into separate class library
Fixes #39
