+++ samples/eShopLite/eShopLite.AppHost/eShopLite.AppHost.csproj
@@ -10,6 +10,7 @@
 
   <ItemGroup>
     <PackageReference Include="Aspire.Hosting" />
+    <ProjectReference Include="..\eShopLite.CatalogDbManager\eShopLite.CatalogDbManager.csproj" />
     <ProjectReference Include="..\eShopLite.CatalogDb\eShopLite.CatalogDb.csproj" />
     <ProjectReference Include="..\eShopLite.Frontend\eShopLite.Frontend.csproj" />
     <ProjectReference Include="..\eShopLite.CatalogService\eShopLite.CatalogService.csproj" />
