+++ samples/eShopLite/eShopLite.CatalogDb/CatalogBrand.cs
@@ -1,0 +1,7 @@
+namespace eShopLite.CatalogDb;
+
+public class CatalogBrand
+{
+    public int Id { get; set; }
+    public required string Brand { get; set; }
+}
