+++ samples/eShopLite/eShopLite.CatalogDb/CatalogDbContext.cs
@@ -3,8 +3,6 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace eShopLite.CatalogDb;
 
-public record Catalog(int FirstId, int NextId, bool IsLastPage, IEnumerable<CatalogItem> Data);
-
 public class CatalogDbContext(DbContextOptions<CatalogDbContext> options) : DbContext(options)
 {
     // https://learn.microsoft.com/ef/core/performance/advanced-performance-topics#compiled-queries
@@ -106,36 +104,4 @@ public class CatalogDbContext(DbContextOptions<CatalogDbContext> options) : DbCo
 
         return results;
     }
-}
-
-public class CatalogType
-{
-    public int Id { get; set; }
-    public required string Type { get; set; }
-}
-
-public class CatalogBrand
-{
-    public int Id { get; set; }
-    public required string Brand { get; set; }
-}
-
-public class CatalogItem
-{
-    public int Id { get; set; }
-    public required string Name { get; set; }
-    public string? Description { get; set; }
-    public decimal Price { get; set; }
-    public required string PictureFileName { get; set; }
-    public string? PictureUri { get; set; }
-
-    public int CatalogTypeId { get; set; }
-    public required CatalogType CatalogType { get; set; }
-
-    public int CatalogBrandId { get; set; }
-    public required CatalogBrand CatalogBrand { get; set; }
-    public int AvailableStock { get; set; }
-    public int RestockThreshold { get; set; }
-    public int MaxStockThreshold { get; set; }
-    public bool OnReorder { get; set; }
 }
