+++ samples/eShopLite/eShopLite.CatalogDb/CatalogType.cs
@@ -1,0 +1,7 @@
+namespace eShopLite.CatalogDb;
+
+public class CatalogType
+{
+    public int Id { get; set; }
+    public required string Type { get; set; }
+}
