+++ samples/eShopLite/eShopLite.CatalogDbManager/Program.cs
@@ -1,9 +1,12 @@
 ﻿using eShopLite.CatalogDb;
+using eShopLite.CatalogDbManager;
+using Microsoft.EntityFrameworkCore;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.AddServiceDefaults();
-builder.AddNpgsqlDbContext<CatalogDbContext>("catalogdb");
+builder.AddNpgsqlDbContext<CatalogDbContext>("catalogdb", null,
+    c => c.UseNpgsql(a => a.MigrationsAssembly(typeof(Program).Assembly.GetName().Name)));
 
 builder.Services.AddOpenTelemetry()
     .WithTracing(tracing => tracing.AddSource(CatalogDbInitializer.ActivitySourceName));
