+++ samples/eShopLite/eShopLite.CatalogDbManager/Properties/launchSettings.json
