+++ samples/eShopLite/eShopLite.CatalogDbManager/appsettings.Development.json
