+++ samples/eShopLite/eShopLite.CatalogDbManager/appsettings.json
