+++ samples/eShopLite/eShopLite.CatalogService/CatalogApi.cs
@@ -52,3 +52,5 @@ public static class CatalogApi
         return group;
     }
 }
+
+public record Catalog(int FirstId, int NextId, bool IsLastPage, IEnumerable<CatalogItem> Data);
